+++ examples/be-example-express-sqlite/src/agent-playground.ts
@@ -1,4 +1,4 @@
-import {playground} from "@convostack/playground";
-import {AgentEcho} from "@convostack/agent-echo";
+import { playground } from "convostack/playground";
+import { AgentEcho } from "convostack/agent-echo";
 
-playground(new AgentEcho())
+playground(new AgentEcho());